some const to int

+++ babylon-helper/src/jsMain/kotlin/casper/model/ModelCreateOptions.kt
@@ -1,3 +1,3 @@
 package casper.model
 
-class ModelCreateOptions(val ignoreCameras: Boolean = true, val ignoresLights: Boolean = true, val isPickable: Boolean? = null, val cullingStrategy: Double? = null, val castShadow: Boolean = true, val receiveShadow: Boolean = true)
+class ModelCreateOptions(val ignoreCameras: Boolean = true, val ignoresLights: Boolean = true, val isPickable: Boolean? = null, val cullingStrategy: Int? = null, val castShadow: Boolean = true, val receiveShadow: Boolean = true)
